requirements.txt
google-cloud-bigquery==3.17.2
pandas==2.2.1
dbt-bigquery==1.7.7
streamlit==1.33.0
anthropic==0.25.0
python-dotenv==1.0.1
db-dtypes==1.2.0
pyarrow==15.0.2
